Testing Git Pages Feature for a certain Repo
This is a test commit to test the behaviour of Github when selecting "Only use /docs Folder for Github Pages" in the Settings of my Doc Repo.
Expected behaviour: It won't find an Index.html (as there is none in the Root of the actul Docs. If that's the case I might need to only add the relevant API Folder "as a root")